Extract progress callbacks
